+++ packages/cli/__tests__/zostso/__unit__/send/Send.definition.unit.test.ts
@@ -15,7 +15,7 @@ describe("zos-tso send group definition", () => {
     it("should not have changed", () => {
         const definition: ICommandDefinition = require("../../../../src/zostso/send/Send.definition").SendCommand;
         expect(definition).toBeDefined();
-        expect(definition.children.length).toBe(1);
+        expect(definition.children.length).toBe(2);
         delete definition.children;
         expect(definition).toMatchSnapshot();
     });