+++ packages/imperative/src/imperative/src/plugins/utilities/npm-interface/install.ts
@@ -28,6 +28,7 @@ import { IProfileTypeConfiguration } from "../../../../../profiles";
 import * as semver from "semver";
 import { ConfigUtils } from "../../../../../config";
 import { IExtendersJsonOpts } from "../../../../../config/src/doc/IExtenderOpts";
+import { INpmRegistryInfo } from "../../doc/INpmRegistryInfo";
 
 // Helper function to update extenders.json object during plugin install.
 // Returns true if the object was updated, and false otherwise
@@ -74,9 +75,7 @@ export const updateExtendersJson = (
  *                                 be converted to an absolute path prior to being passed to the
  *                                 `npm install` command.
  *
- * @param {string} registry The npm registry to use, this is expected to be passed by every caller
- *                          so if calling functions don't have a registry available, they need
- *                          to get it from npm.
+ * @param {INpmRegistryInfo} registryInfo The npm registry to use.
  *
  * @param {boolean} [installFromFile=false] If installing from a file, the package location is
  *                                          automatically interpreted as an absolute location.
@@ -87,7 +86,7 @@ export const updateExtendersJson = (
  *                                          it.
  * @returns {string} The name of the plugin.
  */
-export async function install(packageLocation: string, registry: string, installFromFile = false) {
+export async function install(packageLocation: string, registryInfo: INpmRegistryInfo, installFromFile = false) {
     const iConsole = Logger.getImperativeLogger();
     let npmPackage = packageLocation;
 
@@ -117,12 +116,15 @@ export async function install(packageLocation: string, registry: string, install
     }
 
     try {
-        iConsole.debug(`Installing from registry ${registry}`);
+        iConsole.debug(`Installing from registry ${registryInfo.location}`);
 
         // Perform the npm install.
         iConsole.info("Installing packages...this may take some time.");
 
-        installPackages(PMFConstants.instance.PLUGIN_INSTALL_LOCATION, registry, npmPackage);
+        installPackages(npmPackage, {
+            prefix: PMFConstants.instance.PLUGIN_INSTALL_LOCATION,
+            ...registryInfo.npmArgs,
+        });
 
         // We fetch the package name and version of newly installed plugin
         const packageInfo = await getPackageInfo(npmPackage);
@@ -150,7 +152,7 @@ export async function install(packageLocation: string, registry: string, install
 
         const newPlugin: IPluginJsonObject = {
             package: npmPackage,
-            location: registry,
+            location: registryInfo.location,
             version: packageVersion
         };
         iConsole.debug("Updating the current configuration with new plugin:\n" +